new organization and readme
