+++ client/functions.h
@@ -1,0 +1,12 @@
+#ifndef FUNCTIONS_H_INCLUDED
+#define FUNCTIONS_H_INCLUDED
+#pragma once
+
+
+void catchfunctions(char argv[]);
+
+int countparams(char string[]);
+
+int callfunction(int i, char** params);
+
+#endif // FUNCTIONS_H_INCLUDED
