+++ lib/edit.h
@@ -1,0 +1,13 @@
+#ifndef EDIT_H_INCLUDED
+#define EDIT_H_INCLUDED
+
+typedef struct Edit Edit;
+
+struct Edit
+{
+   Edit* next;
+   char* name;   
+};
+
+
+#endif // EDITOR_H_INCLUDED
